Jessica cccourse routes

+++ routes/student-routes.js
@@ -1,6 +1,7 @@
 'use strict';
 
 // npm modules
+const createError = require('http-errors');
 const Router = require('express').Router;
 
 // app modules
@@ -11,6 +12,8 @@ const bearerAuth = require('../lib/bearer-auth-middleware.js');
 const router = module.exports = new Router();
 
 router.get('/students', bearerAuth, (req, res, next) => {
+  if(!req.student) return next(createError(401));
+
   Student.findById(req.student._id)
     .then(student => {
       delete student.password;
@@ -20,6 +23,8 @@ router.get('/students', bearerAuth, (req, res, next) => {
 });
 
 router.put('/students', bearerAuth, (req, res, next) => {
+  if(!req.student) return next(createError(401));
+
   Student.findOneAndUpdate({_id: req.student._id}, req.body, {new: true})
     .then(student => {
       delete student.password;
@@ -29,6 +34,8 @@ router.put('/students', bearerAuth, (req, res, next) => {
 });
 
 router.delete('/students', bearerAuth, (req, res, next) => {
+  if(!req.student) return next(createError(401));
+
   Student.findByIdAndRemove(req.student._id)
     .then(() => res.status(204).end())
     .catch(next);
